Add sidebar module
